cadastral_proccess/requirements.txt
amqp==5.1.1
asgiref==3.7.2
async-timeout==4.0.3
billiard==4.1.0
celery==5.3.4
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
Django==4.1
djangorestframework==3.14.0
gevent==23.9.1
greenlet==2.0.2
idna==3.4
kombu==5.3.2
packaging==23.1
prompt-toolkit==3.0.39
pycparser==2.21
python-dateutil==2.8.2
pytz==2023.3.post1
redis==5.0.0
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.7.1
tzdata==2023.3
urllib3==2.0.4
vine==5.0.0
wcwidth==0.2.6
zope.event==5.0
zope.interface==6.0
